refactor(api): move utils for routes from sites
